MatrixtechBack/requirements.txt
autopep8==1.6.0
click==8.1.2
Flask==2.1.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
pycodestyle==2.8.0
PyMySQL==1.0.2
pyparsing==3.0.8
six==1.16.0
SQLAlchemy==1.4.35
toml==0.10.2
Werkzeug==2.1.1
zipp==3.8.0
pyjwt==2.3.0
dropbox==11.29.0
pillow==7.0.0
bcrypt==3.2.0